Refactor: move RemoteFolder to feature:folder:api-module
