+++ legacy/core/src/main/java/com/fsck/k9/mailstore/SpecialFolderUpdater.kt
@@ -4,9 +4,9 @@ import app.k9mail.core.common.mail.Protocols
 import app.k9mail.core.mail.folder.api.FolderType
 import app.k9mail.legacy.account.Account
 import app.k9mail.legacy.account.Account.SpecialFolderSelection
-import app.k9mail.legacy.folder.RemoteFolder
 import app.k9mail.legacy.mailstore.FolderRepository
 import com.fsck.k9.Preferences
+import net.thunderbird.feature.folder.api.RemoteFolder
 
 /**
  * Updates special folders in [Account] if they are marked as [SpecialFolderSelection.AUTOMATIC] or if they are marked
