+++ legacy/ui/folder/src/main/java/app/k9mail/legacy/ui/folder/FolderNameFormatter.kt
@@ -3,7 +3,7 @@ package app.k9mail.legacy.ui.folder
 import android.content.res.Resources
 import app.k9mail.core.mail.folder.api.Folder
 import app.k9mail.core.mail.folder.api.FolderType
-import app.k9mail.legacy.folder.RemoteFolder
+import net.thunderbird.feature.folder.api.RemoteFolder
 
 class FolderNameFormatter(private val resources: Resources) {
     fun displayName(folder: Folder): String {
