+++ settings.gradle.kts
@@ -118,6 +118,7 @@ include(
 
 include(
     ":feature:funding:api",
+    ":feature:folder:api",
     ":feature:funding:googleplay",
     ":feature:funding:link",
     ":feature:funding:noop",
